+++ test/core-remote/__mocks__/node-scp.js
@@ -1,3 +1,5 @@
+const mockdata = require('../../mock-remotefilelist')
+
 module.exports = function (scpLoginOption) {
     return new Promise((resolve, reject) => {
         if (expect.getState().currentTestName == 'remote/login/success')
@@ -13,7 +15,7 @@ module.exports.mockClient = {
         return new Promise((resolve, reject) => {
             if (expect.getState().currentTestName == 'remote/getFileList/success' ||
                 expect.getState().currentTestName == 'remote/getFileList/success/quiet') {
-                resolve(mockRemoteFileList);
+                resolve(mockdata.mockRemoteFileList);
             } else if (expect.getState().currentTestName == 'remote/getFileList/fail' ||
                 expect.getState().currentTestName == 'remote/getFileList/fail/quiet'
             )
@@ -23,26 +25,3 @@ module.exports.mockClient = {
         });
     }
 }
-
-mockRemoteFileList = [
-    {
-        type: '-',
-        name: 'Mock_File_1.zip',
-        size: 2928,
-        modifyTime: 1622867586000,
-        accessTime: 1623577546000,
-        rights: { user: 'rw', group: 'rw', other: 'r' },
-        owner: 1001,
-        group: 1001
-    },
-    {
-        type: '-',
-        name: 'Mock_File_2.zip',
-        size: 49453,
-        modifyTime: 1622867586000,
-        accessTime: 1623577546000,
-        rights: { user: 'rw', group: 'rw', other: 'r' },
-        owner: 1001,
-        group: 1001
-    }
-];