+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - `#49`: As dev/test, I want to refactor code and do jest testing. (`scp.login` to `core-remote.login`)
 - `#54`: As dev/test, I want to refactor code and do jest testing. (`scp.getRemoteFileList` to `core-remote.getFileList`)
 - `#55`: As dev/test, I want to refactor code and do jest testing. (`scp.filterFiles` to `corefilters.filterFiles`, `corefilters.setupFilters`)
+- `#73`: As test, I want to combine to mock-data file.
 
 ## `0.7.0`